refactor to models and factory
