+++ react-frontend/src/Models/ScreenshotURL.js
@@ -1,0 +1,14 @@
+/*
+@dataclass
+class ScreenshotURL:
+	thumbnail_url: str
+	fullsize_url: str
+*/
+export default class ScreenshotURL {
+    constructor(
+      thumbnail_url,
+      fullsize_url) {
+        this.thumbnail_url = thumbnail_url
+        this.fullsize_url = fullsize_url
+    }
+}
